db model and internal process refactoring
Refactored both the internal process class and the models class to support foreign keys. It turns out that the field names for foreign keys should not include id, as this is used by django to access the actual underlying id values. Furthermore, each foreign key should specificy a db_column that is the actual name of the field the id will be stored in within the database.