external transfer work
external transfers will now process requests by moving funds out of the internal account and to an external transfer pool. While external transfers are considered complete at this point, they are not actually complete until the ACH transfer goes through. I don't know if we'll have time to actually set up some kind of ACH tranfer feature to other banks, but we'll assume that thi is as far as we need to go. Keep in mind that any of these transfers can be reversed since the amount removed is stored in this table.