autopayments and accounts
I've completed (but not tested) the methods for modifying, canceling, and retrieving autopayments.

I've also added some methods in ExchangeProcessor that allow you to retrieve all exchange records associated with a specific account number; this will be used in the account information retrival function to pull up transfer historys for an account.