account details view
What makes my skin crawl is the fact that we are using the bank account number as the ID for the front end, I really don't like this. I think it would be better if we used the primary key from the database, and then sent that along with the account number to the front end. That way, we can redact/block all but the last 5 digits of the account number and display that on the front end. We can allow the user to view the account number through a modal or something.

Anyway, that's for a later refactor