New Account Number fields
Users will now enter bank account numbers as a character. They will need to enter it as a 9 or 12 digit number (even if it's all just zeroes). I haven't tested this change yet so i'll have to see if it breaks anything. Furthermore, i will need to modify the front end to always display the numbers as a 9/12 digit number