autopayment fix and trasnfer fix
You can no longer setup an automatic payment from an account that doesn't have enough money to make atleast 1 payment.

You also cannot setup a transfer to and from the same account (while this was already prevented on the backend, the front end will now notify the user).

Finally, successful transfers now redirect the user to the account that the money was transferd from (as this will always be accessible to the usre)